Allow embedding user flags in level specifications
Log levels are now 16-bit unsigned integers. The least significant 8
bits constitute the log level (a number in the range [0,256)

The 8 most significant bits are used to encode user flags. If messages and
parameters are logged with a log level containing a user flag, they're
only logged if the same flags are set in the global log level.

This can be used to filter messages based on other criteria than a
single numerical severity scale. For example, you may tag a parameter as
containing patient identifiable information (PII) or other sensitive
data, which should normally be omitted. However, in debugging
environments you may wish to output these parameters regardless.